requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-decorator-include==3.0
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
excel2json-3==0.1.6
flake8==4.0.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.3
openpyxl==3.0.9
packaging==21.0
pandas==1.3.4
Pillow==8.2.0
pipfile==0.0.2
# psycopg2==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.1
pytils==0.3
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
